requirements.txt
# Discord Bot
discord.py>=2.3.2

# Web Scraping
playwright>=1.40.0

# Database
SQLAlchemy>=2.0.23
aiomysql>=0.2.0
pymysql>=1.1.0
alembic>=1.13.0

# Task Scheduling
APScheduler>=3.10.4

# Utilities
python-dotenv>=1.0.0